requirements.txt
selenium
beautifulsoup4
opencv-python
pandas
google-generativeai
easyocr
Pillow
openpyxl